teams_bot/teams_bot/requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
#cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
dialogflow==1.1.0
Django==3.1.4
django-cors-headers==3.7.0
djangorestframework==3.12.2
google-api-core==1.24.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
#grpcio==1.34.0
gunicorn==20.1.0
idna==2.10
#numpy==1.21.2
#pandas==1.3.3
protobuf==3.14.0
#psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2020.5
requests==2.25.1
rsa==4.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
XlsxWriter==1.4.3
